djangoAPI/requirements.txt
Django==5.2.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
